+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This changelog adheres to semantic version according to [keepachangelog](https:/
 
 ## Unreleased
 
+### Added
+- MIME type detection in `--body` flag if file is given.
+
 ### Removed
 - `comp` command